Meta attributes as variables
Updating rrssb.js per issue #47 (https://github.com/kni-labs/rrssb/issues/47) 

Created the publicly accessible function $.rrssb() which accepts description, emailAddress, emailBody, emailSubject, image, title and url as options. 

It also includes an encode function at line 70 for recursively decoding strings and then encoding them a single time.

This should be fully backward compatible, if you don't know it exists nothing will have changed in the function of the plugin but if you want to override the attributes on a case by case basis, this should be a simple way of doing so.